Move flutter_web.dart to project.dart and remove FlutterWebManager class. (#703)
This class contained only static members (and a few accidental instance
members), so does not need to be a class. The methods have grown to handle both
Flutter Web concepts and non-Flutter web concepts, but all members seem to be
concerned with the "project" created in the DartPad script.

So I've moved the file to project.dart and changed all imports to to use the
prefix `project`, since this library primarily exposes top-level functions.
